+++ app/javascript/packs/general/UserMenu.jsx
@@ -14,6 +14,7 @@ const UserMenu = () => {
     function logOut () {
         cookies.remove('Name',{path:'/'});
         cookies.remove('Type',{path:'/'});
+        cookies.remove('Appointment',{path:'/'});
         axios.post("/application/0/log_out", {}, {
             withCredentials: true
         })
@@ -25,15 +26,15 @@ const UserMenu = () => {
       <div className='user-menu'>
         {cookies.get("Type") == "Admin" &&
          <button className="admin--button" onClick={() => {toUrl('/admin')}}>Admin Page</button>}
-        {cookies.get("Type") != "Boy" &&
+        {(cookies.get("Type") != "Boy" || cookies.get("Appointment") != null) &&
          <button className="user-management--button" onClick={() => {toUrl('/user_management')}}>Users Management</button>}
-        {cookies.get("Type") != "Boy" &&
+        {(cookies.get("Type") != "Boy" || cookies.get("Appointment") != null) &&
          <button className="award-management--button" onClick={() => {toUrl('/awards')}}>Awards</button>}
-        {cookies.get("Type") != "Boy" &&
+        {(cookies.get("Type") != "Boy" || cookies.get("Appointment") != null) &&
          <button className="result-generation--button" onClick={() => {toUrl('/generate_result')}}>Result Generation</button>}
         {cookies.get("Type") != "Boy" &&
          <button className="uniform-inspection--button" onClick={() => {toUrl('/uniform_inspection_results')}}>Uniform Inspection</button>}
-        {cookies.get("Type") != "Boy" &&
+        {(cookies.get("Type") != "Boy" || cookies.get("Appointment") != null) &&
          <button className="help--button" onClick={() => {toUrl('/help')}}>Help</button>}
         <button onClick={() => {toUrl('/reset_password')}}>Reset Password</button>
         <button className="log-out--button" onClick={logOut}>Log Out</button>