+++ db/migrate/202411040000001_add_appointment_to_accounts.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class AddAppointmentToAccounts < ActiveRecord::Migration[7.0]
+    def change
+      change_table :accounts do |t|
+        t.string :appointment
+      end
+    end
+  end